Refactor creating DAG to new file
